requirements.txt
streamlit==1.40.1
pydantic==2.7.1
python-dotenv==1.0.0
requests==2.32.3
